refactor(features): moved getSubjectsFromFirestore to features
